화면이 심심한거 같아서 달을 그리고, UFO를 넣기로 했음 달과 UFO클래스를 추가함. 마을 스카이라인을 클래스로 만듦. 이 과정에서 좀더 객체 지향적으로 수정함. 눈이 UFO 위에 그려져야 하는데 생성순서 때문에 UFO에 가져지는 문제가 있어서 Layer를 만들어야함

+++ SnowVillage/Classes/Wind.cs
@@ -24,7 +24,7 @@ namespace SnowVillage
         /// <summary>
         /// 바람의 방향-세기가 바뀌는 주기  nanosecond
         /// </summary>
-        private const int changeWindInterval = 50000000;
+        private const long changeWindInterval = 50000000;
 
         /// <summary>
         /// 바람의 방향-세기 주기 마다 바뀝니다.
